Makes attribute assignment make more sense
Attribute assignment previously was assigning an instance variable to
the result of a method that lazy initialized the same instance variable.
This was causing a bug in an app on the line that this initialization
was performed. This change makes the code a little easier to understand
and fixes the bug.